+++ tests/run_tests.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 '''VIC testing command line interface'''
 
-from __future__ import print_function
 import os
 import sys
 import glob
@@ -9,22 +8,19 @@ import argparse
 import datetime
 from collections import OrderedDict
 import string
-import xarray as xr
 
 import pytest
-import pandas as pd
-import numpy as np
 
-from tonic.models.vic.vic import (VIC, VICRuntimeError,
-                                  default_vic_valgrind_error_code)
+from tonic.models.vic.vic import VIC, default_vic_valgrind_error_code
 from tonic.io import read_config, read_configobj
-from tonic.testing import check_completed, check_for_nans, VICTestError
-from test_image_driver import assert_nan_equal
+from tonic.testing import VICTestError
+from test_utils import *
+from test_image_driver import *
+from test_restart import *
 
 test_dir = os.path.dirname(os.path.abspath(__file__))
 
 OUTPUT_WIDTH = 100
-ERROR_TAIL = 20  # lines
 
 description = '''
                             VIC Test Suite
@@ -84,15 +80,6 @@ class TestResults(object):
         return '{0} test results:Passed: {1}, Comment:{2}'.format(self.name,
                                                                   self.passed,
                                                                   self.comment)
-
-
-class VICReturnCodeError(Exception):
-    pass
-
-
-class VICValgrindError(Exception):
-    pass
-
 
 def main():
     '''
@@ -735,594 +722,6 @@ def run_release(config_file):
     '''
     return OrderedDict()
 
-
-def setup_test_dirs(testname, out_dir, mkdirs=['results', 'state',
-                                               'logs', 'plots']):
-    '''create test directories for testname'''
-
-    dirs = OrderedDict()
-    dirs['test'] = os.path.join(out_dir, testname)
-    for d in mkdirs:
-        dirs[d] = os.path.join(dirs['test'], d)
-
-    for dirname in dirs.values():
-        os.makedirs(dirname, exist_ok=True)
-
-    return dirs
-
-
-def print_test_dict(d):
-    '''print a nicely formatted set of test results'''
-    print('{0: <48} | {1: <6} | {2}'.format('Test Name', 'Passed', 'Comment'))
-    print('-'.ljust(OUTPUT_WIDTH, '-'))
-    for k, v in d.items():
-        print('{0: <48} | {1: <6} | {2}'.format(v.name, str(bool(v.passed)),
-                                                v.comment))
-        print('-'.ljust(OUTPUT_WIDTH, '-'))
-
-
-def print_tail(string, n=20, indent='\t--->'):
-    '''print tail of multiline string'''
-    lines = string.decode().splitlines()
-    for l in lines[-n:]:
-        print('{0}{1}'.format(indent, l))
-
-
-def replace_global_values(gp, replace):
-    '''given a multiline string that represents a VIC global parameter file,
-       loop through the string, replacing values with those found in the
-       replace dictionary'''
-    gpl = []
-    for line in iter(gp.splitlines()):
-        line_list = line.split()
-        if line_list:
-            key = line_list[0]
-            if key in replace:
-                value = replace.pop(key)
-                val = list([str(value)])
-            else:
-                val = line_list[1:]
-            gpl.append('{0: <20} {1}\n'.format(key, ' '.join(val)))
-
-    if replace:
-        for key, val in replace.items():
-            try:
-                value = ' '.join(val)
-            except:
-                value = val
-            gpl.append('{0: <20} {1}\n'.format(key, value))
-
-    return gpl
-
-
-def drop_tests(config, driver):
-    '''helper function to remove tests that should not be run for driver'''
-    new = {}
-    for key, test_cfg in config.items():
-        try:
-            if test_cfg['driver'].lower() == driver.lower():
-                new[key] = test_cfg
-        except KeyError:
-            raise KeyError('test configuration must specify driver')
-    return new
-
-
-def pop_run_kwargs(config):
-    '''pop run kwargs for VIC executable'''
-    run_kwargs = {}
-    run_kwargs['valgrind'] = config.pop('valgrind', False)
-    run_kwargs['mpi_proc'] = config.pop('mpi_proc', None)
-    return run_kwargs
-
-
-def check_returncode(returncode, expected=0):
-    '''check return code given by VIC, raise error if appropriate'''
-    if returncode == expected:
-        return None
-    elif returncode == default_vic_valgrind_error_code:
-        raise VICValgrindError('Valgrind raised an error')
-    else:
-        raise VICReturnCodeError('VIC return code ({0}) does not match '
-                                 'expected ({1})'.format(returncode, expected))
-
-
-def process_error(error, vic_exe):
-    '''Helper function to process possible error raised during testing'''
-    if isinstance(error, VICRuntimeError):
-        test_comment = 'Test failed during simulation'
-        error_message = error
-        tail = vic_exe.stderr
-    elif isinstance(error, VICTestError):
-        test_comment = 'Test failed during testing of output files'
-        error_message = error
-        tail = None
-    elif isinstance(error, VICValgrindError):
-        test_comment = 'Test failed due to memory error detected by valgrind'
-        error_message = error
-        tail = vic_exe.stderr
-    elif isinstance(error, VICReturnCodeError):
-        test_comment = 'Test failed due to incorrect return code'
-        error_message = error
-        tail = vic_exe.stderr
-    else:
-        raise error
-
-    print('\t{0}'.format(test_comment))
-    print('\t{0}'.format(error_message))
-    if tail is not None:
-        print('\tLast {0} lines of standard out:'.format(ERROR_TAIL))
-        print_tail(tail, n=ERROR_TAIL)
-
-    return test_comment, error_message
-
-
-def test_classic_driver_all_complete(fnames):
-    '''
-    Test that all VIC files in fnames have the same first and last index position
-    '''
-    start = None
-    end = None
-    for fname in fnames:
-        df = read_vic_ascii(fname, header=True)
-
-        # check that each dataframe includes all timestamps
-        if (start is not None) and (end is not None):
-            check_completed(df, start, end)
-        else:
-            start = df.index[0]
-            end = df.index[-1]
-
-
-def test_classic_driver_no_output_file_nans(fnames):
-    '''Test that all VIC classic driver output files in fnames have no nans'''
-    for fname in fnames:
-        df = read_vic_ascii(fname, header=True)
-        check_for_nans(df)
-
-
-def test_image_driver_no_output_file_nans(fnames, domain_file):
-    '''
-    Test that all VIC image driver output files have the same nan structure as
-    the domain file
-    '''
-    for fname in fnames:
-        ds_domain = xr.open_dataset(domain_file)
-        ds_output = xr.open_dataset(fname)
-        assert_nan_equal(ds_domain, ds_output)
-
-
-# TODO: Update tonic version of this function, need to check that subdaily works
-def read_vic_ascii(filepath, header=True, parse_dates=True,
-                   datetime_index=None, names=None, **kwargs):
-    '''Generic reader function for VIC ASCII output with a standard header
-    filepath: path to VIC output file
-    header (True or False):  Standard VIC header is present
-    parse_dates (True or False): Parse dates from file
-    datetime_index (Pandas.tseries.index.DatetimeIndex):  Index to use as
-    datetime index names (list like): variable names
-    **kwargs: passed to Pandas.read_table
-    returns Pandas.DataFrame
-    '''
-    kwargs['header'] = None
-
-    if header:
-        kwargs['skiprows'] = 4
-
-        # get names
-        if names is None:
-            with open(filepath) as f:
-                # skip lines 0 through 3
-                for _ in range(3):
-                    next(f)
-
-                # process header
-                names = next(f)
-                names = names.strip('#').replace('OUT_', '').split()
-
-    kwargs['names'] = names
-
-    if parse_dates:
-        time_cols = ['YEAR', 'MONTH', 'DAY']
-        if 'SECONDS' in names:
-            time_cols.append('SECONDS')
-        kwargs['parse_dates'] = {'datetime': time_cols}
-        kwargs['index_col'] = 0
-
-    df = pd.read_table(filepath, **kwargs)
-
-    if datetime_index is not None:
-        df.index = datetime_index
-
-    return df
-
-
-def prepare_restart_run_periods(restart_dict, state_basedir, statesec):
-    ''' For restart tests, read full running period and splitting dates into datetime objects
-    Parameters
-    ----------
-    restart_dict: <class 'configobj.Section'>
-        A section of the config file for exact restart test setup
-    state_basedir: <str>
-        Basedirectory of output state files.
-        For restart tests, state files will be output as <state_basedir>/<run_start_date>_<run_end_date>/<state_file>
-    statesec: <int>
-        STATESEC option in global parameter file
-
-    Returns
-    ----------
-    run_periods: OrderedDict
-        A list of running periods, including the full-period run, and all splitted 
-        runs in order. Each element of run_period is a dictionary with keys:
-            start_date
-            end_date
-            init_state  # None, or full path of the initial state file
-                        # e.g., '/path/19490101_19490105/states_19490105_82800'
-        
-    '''
-
-    #--- Read in full running period ---#
-    start_date = datetime.datetime(restart_dict['start_date'][0],
-                                   restart_dict['start_date'][1],
-                                   restart_dict['start_date'][2])
-    end_date = datetime.datetime(restart_dict['end_date'][0],
-                                 restart_dict['end_date'][1],
-                                 restart_dict['end_date'][2])
-    #--- Identify each of the splitted running period ---#
-    list_split_dates = []
-    for i, split in enumerate(restart_dict['split_dates']):
-        list_split_dates.append(datetime.datetime(restart_dict['split_dates'][split][0],
-                                                  restart_dict['split_dates'][split][1],
-                                                  restart_dict['split_dates'][split][2]))
-    #--- Prepare running periods ---#
-    # run_periods is a list of running periods, including the full-period run, 
-    # and all splitted runs in order. Each element of run_period is a dictionary 
-    # with keys:
-    #       start_date
-    #       end_date
-    #       init_state  # None, or full path of the initial state file
-    #                   # e.g., '/path/19490101_19490105/states_19490105_82800'
-    run_periods = []
-    # Append the full run
-    d = dict(start_date=start_date, end_date=end_date)
-    d['init_state'] = None
-    run_periods.append(d)
-    # First splitted running period - start_date to first split date
-    d = dict(start_date=start_date, end_date=list_split_dates[0])
-    d['init_state'] = None
-    run_periods.append(d)
-    # Loop over each of the rest splitted periods
-    for i in range(len(list_split_dates)-1):
-        d = dict(start_date=list_split_dates[i] + dt.timedelta(days=1),
-                 end_date=list_split_dates[i+1])
-        d['init_state'] = os.path.join(
-                state_basedir,
-                '{}_{}'.format(run_periods[-1]['start_date'].strftime("%Y%m%d"),
-                               run_periods[-1]['end_date'].strftime("%Y%m%d")),
-                '{}{}_{}'.format('states_',
-                                 run_periods[-1]['end_date'].strftime("%Y%m%d"),
-                                 statesec))
-    run_periods.append(d)
-    # Last splitted running period - last split date to end_date
-    d = dict(start_date=list_split_dates[len(list_split_dates)-1] + 
-                        datetime.timedelta(days=1),
-             end_date=end_date)
-    d['init_state'] = os.path.join(
-            state_basedir,
-            '{}_{}'.format(run_periods[-1]['start_date'].strftime("%Y%m%d"),
-                           run_periods[-1]['end_date'].strftime("%Y%m%d")),
-            '{}{}_{}'.format('states_',
-                             run_periods[-1]['end_date'].strftime("%Y%m%d"),
-                             statesec))
-    run_periods.append(d)
-
-    return run_periods
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def find_global_param_value(gp, param_name):
-    ''' Return the value of a global parameter
-
-    Parameters
-    ----------
-    gp: <str>
-        Global parameter file, read in by read()
-    param_name: <str>
-        The name of the global parameter to find
-
-    Returns
-    ----------
-    line_list[1]: <str>
-        The value of the global parameter
-    '''
-    for line in iter(gp.splitlines()):
-        line_list = line.split()
-        if line_list==[]:
-            continue
-        key = line_list[0]
-        if key==param_name:
-            return line_list[1]
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def setup_subdirs_restart_test(result_basedir, state_basedir, run_periods):
-    ''' Set up subdirectories for multiple runs for restart testing
-    Parameters
-    ----------
-    result_basedir: <str>
-        Base directory of output fluxes results; running periods are subdirectories under the base directory
-    state_basedir: <str>
-        Base directory of output state results; running periods are subdirectories under the base directory
-    run_periods: <list>
-        A list of running periods. Return from prepare_restart_run_periods()
-
-    Returns
-    ----------
-    None
-    '''
-
-    for j, run_period in enumerate(run_periods):
-        # Set up subdirectories for results and states
-        run_start_date = run_period['start_date']
-        run_end_date = run_period['end_date']
-        result_dir = os.path.join(result_basedir,
-                                  '{}_{}'.format(run_start_date.strftime("%Y%m%d"),
-                                                 run_end_date.strftime("%Y%m%d")))
-        state_dir = os.path.join(state_basedir,
-                                 '{}_{}'.format(run_start_date.strftime("%Y%m%d"),
-                                                run_end_date.strftime("%Y%m%d")))
-        os.makedirs(result_dir, exist_ok=True)
-        os.makedirs(state_dir, exist_ok=True)
-
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def setup_subdirs_and_fill_in_global_param_restart_test(s, run_periods, driver, result_basedir, state_basedir, test_data_dir):
-    ''' Fill in global parameter options for multiple runs for restart testing
-
-    Parameters
-    ----------
-    s: <string.Template>
-        Template of the global param file to be filled in
-    run_periods: <list>
-        A list of running periods. Return from prepare_restart_run_periods()
-    driver: <str>
-        'classic' or 'image'
-    result_basedir: <str>
-        Base directory of output fluxes results; running periods are subdirectories under the base directory
-    state_basedir: <str>
-        Base directory of output state results; running periods are subdirectories under the base directory
-    test_data_dir: <str>
-        Base directory of test data
-
-    Returns
-    ----------
-    '''
-
-    list_global_param = []
-    for j, run_period in enumerate(run_periods):
-        # Set up subdirectories for results and states
-        run_start_date = run_period['start_date']
-        run_end_date = run_period['end_date']
-        result_dir = os.path.join(result_basedir,
-                                  '{}_{}'.format(run_start_date.strftime("%Y%m%d"),
-                                                 run_end_date.strftime("%Y%m%d")))
-        state_dir = os.path.join(state_basedir,
-                                 '{}_{}'.format(run_start_date.strftime("%Y%m%d"),
-                                                run_end_date.strftime("%Y%m%d")))
-        os.makedirs(result_dir, exist_ok=True)
-        os.makedirs(state_dir, exist_ok=True)
-        # Determine initial state
-        if run_period['init_state']==None: # if no initial state
-            init_state='#INIT_STATE'
-        else: # else, the initial state is the last time step
-            init_state='INIT_STATE {}'.format(run_period['init_state'])
-            if driver=='image': # In image driver, the name of
-                                # the state file is 'basepath.*'
-                                # instead of 'basepath_*', and
-                                # ends with ".nc"
-                init_state = init_state.replace("states_", "states.") + '.nc'
-
-        # Fill in global parameter options
-        list_global_param.append(s.safe_substitute(
-                test_data_dir=test_data_dir,
-                result_dir=result_dir,
-                state_dir=state_dir,
-                startyear=run_start_date.year,
-                startmonth=run_start_date.month,
-                startday=run_start_date.day,
-                endyear=run_end_date.year,
-                endmonth=run_end_date.month,
-                endday=run_end_date.day,
-                init_state=init_state,
-                stateyear=run_end_date.year,
-                statemonth=run_end_date.month,
-                stateday=run_end_date.day))
-    return(list_global_param)
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def check_exact_restart_fluxes(result_basedir, driver, run_periods):
-    ''' Checks whether all the fluxes are the same w/ or w/o restart
-
-    Parameters
-    ----------
-    result_basedir: <str>
-        Base directory of output fluxes results; running periods are subdirectories under the base directory
-    driver: <str>
-        'classic' or 'image'
-    run_periods: <list>
-        A list of running periods. Return from prepare_restart_run_periods()
-
-    Returns
-    ----------
-    None
-    '''
-
-    #--- Extract full run period ---#
-    run_full_start_date = run_periods[0]['start_date']
-    run_full_end_date = run_periods[0]['end_date']
-    #--- Read full run fluxes ---#
-    if driver=='classic':
-        result_dir = os.path.join(
-            result_basedir,
-            '{}_{}'.format(run_full_start_date.strftime('%Y%m%d'),
-                           run_full_end_date.strftime('%Y%m%d')))
-        # Read in each of the output flux files
-        dict_df_full_run = {}  # a dict of flux at each grid cell, keyed by flux basename
-        for fname in glob.glob(os.path.join(result_dir, '*')):
-            df = read_vic_ascii(fname, header=True)
-            dict_df_full_run[os.path.basename(fname)] = df
-
-    #--- Loop over the result of each split run period ---#
-    for i, run_period in enumerate(run_periods):
-        # Skip the full run
-        if i==0:
-            continue
-        # Extract running period
-        start_date = run_period['start_date']
-        end_date = run_period['end_date']
-        # Loop over each of the output flux files
-        result_dir = os.path.join(
-            result_basedir,
-            '{}_{}'.format(start_date.strftime('%Y%m%d'),
-                           end_date.strftime('%Y%m%d')))
-        if driver=='classic':
-            for flux_basename in dict_df_full_run.keys():
-                # Read in flux data
-                fname = os.path.join(result_dir, flux_basename)
-                df = read_vic_ascii(fname, header=True)
-                # Extract the same period from the full run
-                df_full_run_split_period = dict_df_full_run[flux_basename].truncate(
-                                                         before=start_date,
-                                                         after=end_date)
-                # Compare split run fluxes with full run
-                df_diff = df - df_full_run_split_period
-                if np.absolute(df_diff).max().max() > pow(10, -6):
-                    raise VICTestError('Restart causes inexact flux outputs '
-                                       'for running period {} - {} at grid cell {}!'.
-                                format(start_date.strftime('%Y%m%d'),
-                                       end_date.strftime('%Y%m%d'),
-                                       flux_basename))
-                else:
-                    continue
-    return
-
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def check_exact_restart_states(state_basedir, driver, run_periods, statesec, state_format='ASCII'):
-    ''' Checks whether all the states are the same w/ or w/o restart.
-        Only test the state at the last time step.
-
-    Parameters
-    ----------
-    state_basedir: <str>
-        Base directory of output state results; running periods are subdirectories under the base directory
-    driver: <str>
-        'classic' or 'image'
-    run_periods: <list>
-        A list of running periods. Return from prepare_restart_run_periods()
-    statesec: <int>
-        STATESEC option in global parameter file
-    state_format: <str>
-        state file format, 'ASCII' or 'BINARY'; only need to specify when driver=='classic'
-
-    Returns
-    ----------
-    None
-    '''
-
-    #--- Read the state at the end of the full run ---#
-    # Extract full run period
-    run_full_start_date = run_periods[0]['start_date']
-    run_full_end_date = run_periods[0]['end_date']
-    # Read the state file
-    if driver=='classic':
-        state_fname = os.path.join(
-            state_basedir,
-            '{}_{}'.format(run_full_start_date.strftime('%Y%m%d'),
-                           run_full_end_date.strftime('%Y%m%d')),
-            'states_{}_{}'.format(run_full_end_date.strftime('%Y%m%d'), statesec))
-        if state_format=='ASCII':
-            states_full_run = read_ascii_state(state_fname)
-        elif state_format=='BINARY':
-            states_full_run = read_binary_state(state_fname)
-
-    #--- Read the state at the end of the last period of run ---#
-    # Extract the last split run period
-    run_last_period_start_date = run_periods[-1]['start_date']
-    run_last_period_end_date = run_periods[-1]['end_date']
-    # Read the state file
-    if driver=='classic':
-        state_fname = os.path.join(
-            state_basedir,
-            '{}_{}'.format(run_last_period_start_date.strftime('%Y%m%d'),
-                           run_last_period_end_date.strftime('%Y%m%d')),
-            'states_{}_{}'.format(run_last_period_end_date.strftime('%Y%m%d'), statesec))
-        if state_format=='ASCII':
-            states = read_ascii_state(state_fname)
-        elif state_format=='BINARY':
-            states = read_binary_state(state_fname)
-
-    #--- Compare split run states with full run ---#
-    if driver=='classic':
-        # If ASCII state file, check if almost the same
-        if state_format=='ASCII':
-            states_diff = states - states_full_run
-            if np.absolute(states_diff).max() > pow(10, -6):
-                raise VICTestError('Restart causes inexact state outputs!')
-            else:
-                return
-        # If BINARY state file, check if exactly the same
-        elif state_format=='BINARY':
-            if states!=states_full_run:
-                raise VICTestError('Restart causes inexact state outputs!')
-            else:
-                return
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def read_ascii_state(state_fname):
-    ''' Read in ascii format state file and convert to a list of numbers
-
-    Parameters
-    ----------
-    state_fname: <str>
-        Path of the state file to be read
-
-    Returns
-    ----------
-    states: <np.array>
-        A np.array of float numbers of the state file
-    '''
-
-    with open(state_fname, 'r') as f:
-        list_states = f.read().split()
-        for i, item in enumerate(list_states):
-            list_states[i] = float(item)
-    return np.asarray(list_states)
-# -------------------------------------------------------------------- #
-
-# -------------------------------------------------------------------- #
-def read_binary_state(state_fname):
-    ''' Read in ascii format state file and convert to a list of numbers
-
-    Parameters
-    ----------
-    state_fname: <str>
-        Path of the state file to be read
-
-    Returns
-    ----------
-    states: <bytes>
-        The full binary state file content
-    '''
-
-    with open(state_fname, 'rb') as f:
-        states = f.read()
-    return states
-
-# -------------------------------------------------------------------- #
 
 if __name__ == '__main__':
     main()
